@@ -3,7 +3,7 @@ package hr.unidu.kz.primjeribiraca;
 import android.app.DatePickerDialog;
 import android.app.Dialog;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
+import androidx.fragment.app.DialogFragment;
 import android.widget.DatePicker;
 import android.widget.TextView;
 
